remove dependency on Sub::Identify
Sub::Util is in core and provides the function needed. When that is not
available, include an inlined PP version. It would be possible to try to
pick alternate implementations, and have dynamic prereqs. For this case,
it does not seem worth the effort.

The primary user of this dist is Moose, which has removed its dependency
on Sub::Identify.